librarywalker/requirements.txt
openai>=1.3.0
tiktoken>=0.5.0
numpy>=1.24.0
faiss-cpu>=1.7.0
python-dotenv>=1.0.0
requests>=2.28.0
PyPDF2>=3.0.0